ADD : 장바구니, 제품 Controller 예외 처리 추가

+++ controllers/cartController.js
@@ -1,9 +1,21 @@
-const cartService = require('../services/cartService');
 const jwt = require('jsonwebtoken');
-const SECRET_KEY = process.env.SECRET_KEY;
+const cartService = require('../services/cartService');
+const userService = require('../services/userService');
 
 const getCurrentCart = async (req, res) => {
     try {
+        const token = await req.headers.authorization;
+
+        if (!token) {
+            throw await res
+                .status(200)
+                .json({ message: 'NEED_TO_LOGIN', currentCart: [] });
+        }
+
+        const { email } = jwt.verify(token, process.env.SECRET_KEY);
+        const findUser = await userService.getUserInfo(email);
+        userId = findUser[0].id;
+
         const currentCart = await cartService.getCurrentCart(userId);
 
         return res.status(200).json(currentCart);
@@ -17,6 +29,12 @@ const updateCart = async (req, res) => {
     try {
         const { productId } = req.params;
         const { quantity, setQuantity } = req.query;
+
+        if (!productId || !quantity || !setQuantity) {
+            return res.status(400).json({
+                message: 'MISSING_PRODUCTID_OR_QUANTITY_OR_SETQUANTITY',
+            });
+        }
 
         const itemsInCart = await cartService.updateCart(
             productId,
@@ -35,6 +53,10 @@ const updateCart = async (req, res) => {
 const deleteItemFromCart = async (req, res) => {
     try {
         const { productId } = req.params;
+
+        if (!productId) {
+            return res.status(400).json({ message: 'MISSING_PRODUCTID' });
+        }
 
         const itemsInCart = await cartService.deleteItemFromCart(
             productId,
